# openapi.json
{"openapi":"3.1.0","info":{"title":"Portalku API","description":"API portal informasi akademik - latihan MK Pemrograman Web Lanjut, D3 Teknologi Informasi PSDKU Lumajang","contact":{"name":"D3 TI PSDKU Lumajang"},"version":"v1"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Pengumuman","description":"Pengelolaan pengumuman akademik"}],"paths":{"/api/pengumuman":{"get":{"tags":["Pengumuman"],"summary":"Daftar seluruh pengumuman","description":"Mengembalikan bentuk ringkas tanpa isi lengkap pengumuman","operationId":"semua","responses":{"200":{"description":"Daftar berhasil diambil","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PengumumanRingkasResponse"}}}}}}},"post":{"tags":["Pengumuman"],"summary":"Menambah pengumuman baru","operationId":"tambah","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PengumumanRequest"}}},"required":true},"responses":{"201":{"description":"Pengumuman berhasil dibuat","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PengumumanResponse"}}}},"422":{"description":"Data tidak lolos validasi","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"409":{"description":"Judul sudah dipakai","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}}}}},"/api/pengumuman/{id}/lampiran":{"get":{"tags":["Pengumuman"],"summary":"Dapatkan daftar lampiran dari suatu pengumuman","operationId":"daftarLampiran","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/LampiranResponse"}}}}}}},"post":{"tags":["Pengumuman"],"summary":"Tambah lampiran baru pada suatu pengumuman","operationId":"tambahLampiran","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LampiranRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LampiranResponse"}}}}}}},"/api/jurnal":{"get":{"tags":["jurnal-controller"],"operationId":"semua_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Jurnal"}}}}}}},"post":{"tags":["jurnal-controller"],"operationId":"tambah_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Jurnal"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Jurnal"}}}}}}},"/api/pengumuman/{id}/dilihat":{"patch":{"tags":["Pengumuman"],"summary":"Tambah jumlah dilihat pengumuman sebanyak 1","operationId":"incrementDilihat","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"204":{"description":"No Content"}}}},"/api/percobaan":{"get":{"tags":["percobaan-controller"],"operationId":"semua_2","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Pengumuman"}}}}}}}},"/api/jurnal/{id}":{"get":{"tags":["jurnal-controller"],"operationId":"satu","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Jurnal"}}}}}},"delete":{"tags":["jurnal-controller"],"operationId":"hapus","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}},"/api/info":{"get":{"tags":["info-controller"],"operationId":"info","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/halo":{"get":{"tags":["halo-controller"],"operationId":"halo","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"PengumumanRequest":{"type":"object","properties":{"judul":{"type":"string","description":"Judul pengumuman, harus unik","example":"Jadwal UTS Semester Genap 2026/2027","maxLength":150,"minLength":0},"isi":{"type":"string","description":"Isi detail pengumuman","example":"Pelaksanaan UTS akan dimulai pada tanggal...","minLength":1},"tanggalTerbit":{"type":"string","format":"date","description":"Tanggal pengumuman diterbitkan","example":"2026-09-22"},"idKategori":{"type":"integer","format":"int32","description":"Nomor kategori","example":1}},"required":["idKategori","tanggalTerbit"]},"ProblemDetail":{"type":"object","properties":{"type":{"type":"string","format":"uri"},"title":{"type":"string"},"status":{"type":"integer","format":"int32"},"detail":{"type":"string"},"instance":{"type":"string","format":"uri"},"properties":{"type":"object","additionalProperties":{"type":"object"}}}},"PengumumanResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"judul":{"type":"string"},"isi":{"type":"string"},"tanggalTerbit":{"type":"string","format":"date"},"jumlahDilihat":{"type":"integer","format":"int32"},"namaKategori":{"type":"string"},"jumlahLampiran":{"type":"integer","format":"int32"}}},"LampiranRequest":{"type":"object","properties":{"namaBerkas":{"type":"string","description":"Nama berkas lampiran","example":"Panduan_Akademik.pdf","minLength":1},"ukuran":{"type":"integer","format":"int32","description":"Ukuran berkas dalam bytes","example":1048576}},"required":["ukuran"]},"LampiranResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int32","example":1},"namaBerkas":{"type":"string","example":"Panduan_Akademik.pdf"},"ukuran":{"type":"integer","format":"int32","example":1048576}}},"Jurnal":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"judul":{"type":"string"},"penerbit":{"type":"string"},"tahunTerbit":{"type":"integer","format":"int32"}}},"Dosen":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"nip":{"type":"string"},"nama":{"type":"string"},"mataKuliah":{"type":"array","items":{"$ref":"#/components/schemas/MataKuliah"},"uniqueItems":true}}},"Kategori":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"nama":{"type":"string"}}},"Kurikulum":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"kode":{"type":"string"},"tahunBerlaku":{"type":"integer","format":"int32"},"status":{"type":"string","enum":["AKTIF","NONAKTIF"]},"prodi":{"$ref":"#/components/schemas/Prodi"},"daftarMataKuliah":{"type":"array","items":{"$ref":"#/components/schemas/MataKuliah"}}}},"Lampiran":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"namaBerkas":{"type":"string"},"ukuran":{"type":"integer","format":"int32"},"pengumuman":{"$ref":"#/components/schemas/Pengumuman"}}},"MataKuliah":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"kode":{"type":"string"},"nama":{"type":"string"},"sksTeori":{"type":"integer","format":"int32"},"sksPraktikum":{"type":"integer","format":"int32"},"kurikulum":{"$ref":"#/components/schemas/Kurikulum"},"pengampu":{"type":"array","items":{"$ref":"#/components/schemas/Dosen"},"uniqueItems":true}}},"Pengumuman":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"judul":{"type":"string"},"isi":{"type":"string"},"tanggalTerbit":{"type":"string","format":"date"},"jumlahDilihat":{"type":"integer","format":"int32"},"kategori":{"$ref":"#/components/schemas/Kategori"},"daftarLampiran":{"type":"array","items":{"$ref":"#/components/schemas/Lampiran"}},"prodiDituju":{"type":"array","items":{"$ref":"#/components/schemas/Prodi"},"uniqueItems":true}}},"Prodi":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"kode":{"type":"string"},"nama":{"type":"string"},"jenjang":{"type":"string","enum":["D3","D4"]},"daftarKurikulum":{"type":"array","items":{"$ref":"#/components/schemas/Kurikulum"}},"pengumuman":{"type":"array","items":{"$ref":"#/components/schemas/Pengumuman"},"uniqueItems":true}}},"PengumumanRingkasResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"judul":{"type":"string"},"namaKategori":{"type":"string"},"tanggalTerbit":{"type":"string","format":"date"},"jumlahDilihat":{"type":"integer","format":"int32"}}}}}}